+++ integration-test/langfuse-integration-examples.spec.ts
@@ -17,17 +17,23 @@ describe("Test Node Example", () => {
 describe("Test Web Example", () => {
   it("builds", () => {
     const cwd = join(__dirname, "..", "examples", "example-web");
+    const resultDependencies = spawnSync("npm", ["run", "build:dependencies"], {
+      cwd,
+      encoding: "utf-8",
+    });
+
+    expect(resultDependencies.status).toBe(0);
+
     const result = spawnSync("npm", ["run", "build"], {
       cwd,
       encoding: "utf-8",
     });
 
-    expect(result.stdout).toContain("Done in");
     expect(result.status).toBe(0);
   });
 });
 
-describe.only("Test Web Next.js Example", () => {
+describe("Test Web Next.js Example", () => {
   it("builds", () => {
     const cwd = join(__dirname, "..", "examples", "example-web-nextjs");
     const resultDependencies = spawnSync("npm", ["run", "build:dependencies"], {
