+++ langfuse-core/src/index.ts
@@ -1,0 +1,688 @@
+import {
+  type LangfuseFetchOptions,
+  type LangfuseFetchResponse,
+  type LangfuseQueueItem,
+  type LangfuseCoreOptions,
+  LangfusePersistedProperty,
+  type CreateLangfuseTraceBody,
+  type LangfuseObject,
+  LangfusePostApiRoutes,
+  type CreateLangfuseEventBody,
+  type CreateLangfuseSpanBody,
+  type CreateLangfuseGenerationBody,
+  type CreateLangfuseScoreBody,
+  type UpdateLangfuseSpanBody,
+  type UpdateLangfuseGenerationBody,
+  type GetLangfuseDatasetParams,
+  type GetLangfuseDatasetResponse,
+  type CreateLangfuseDatasetRunItemBody,
+  type CreateLangfuseDatasetRunItemResponse,
+  type CreateLangfuseDatasetBody,
+  type CreateLangfuseDatasetResponse,
+  type CreateLangfuseDatasetItemBody,
+  type CreateLangfuseDatasetItemResponse,
+  type GetLangfuseDatasetRunResponse,
+  type GetLangfuseDatasetRunParams,
+  type DeferRuntime,
+} from "./types";
+import {
+  assert,
+  generateUUID,
+  removeTrailingSlash,
+  retriable,
+  type RetriableOptions,
+  safeSetTimeout,
+  getEnv,
+} from "./utils";
+export * as utils from "./utils";
+import { SimpleEventEmitter } from "./eventemitter";
+import { getCommonReleaseEnvs } from "./release-env";
+export { LangfuseMemoryStorage } from "./storage-memory";
+
+class LangfuseFetchHttpError extends Error {
+  name = "LangfuseFetchHttpError";
+
+  constructor(public response: LangfuseFetchResponse) {
+    super("HTTP error while fetching Langfuse: " + response.status);
+  }
+}
+
+class LangfuseFetchNetworkError extends Error {
+  name = "LangfuseFetchNetworkError";
+
+  constructor(public error: unknown) {
+    super("Network error while fetching Langfuse", error instanceof Error ? { cause: error } : {});
+  }
+}
+
+function isLangfuseFetchError(err: any): boolean {
+  return typeof err === "object" && (err.name === "LangfuseFetchHttpError" || err.name === "LangfuseFetchNetworkError");
+}
+
+abstract class LangfuseCoreStateless {
+  // options
+  private secretKey: string | undefined;
+  private publicKey: string;
+  baseUrl: string;
+  private flushAt: number;
+  private flushInterval: number;
+  private requestTimeout: number;
+  private removeDebugCallback?: () => void;
+  private debugMode: boolean = false;
+  private pendingPromises: Record<string, Promise<any>> = {};
+  private release: string | undefined;
+
+  // internal
+  protected _events = new SimpleEventEmitter();
+  protected _flushTimer?: any;
+  protected _retryOptions: RetriableOptions;
+
+  // Abstract methods to be overridden by implementations
+  abstract fetch(url: string, options: LangfuseFetchOptions): Promise<LangfuseFetchResponse>;
+  abstract getLibraryId(): string;
+  abstract getLibraryVersion(): string;
+
+  // This is our abstracted storage. Each implementation should handle its own
+  abstract getPersistedProperty<T>(key: LangfusePersistedProperty): T | undefined;
+  abstract setPersistedProperty<T>(key: LangfusePersistedProperty, value: T | null): void;
+
+  constructor(params: { publicKey: string; secretKey?: string } & LangfuseCoreOptions) {
+    const { publicKey, secretKey, ...options } = params;
+    assert(publicKey, "You must pass your Langfuse project's api public key.");
+
+    this.publicKey = publicKey;
+    this.secretKey = secretKey;
+    this.baseUrl = removeTrailingSlash(options?.baseUrl || "https://cloud.langfuse.com");
+    this.flushAt = options?.flushAt ? Math.max(options?.flushAt, 1) : 1;
+    this.flushInterval = options?.flushInterval ?? 10000;
+    this.release = options?.release ?? getEnv("LANGFUSE_RELEASE") ?? getCommonReleaseEnvs() ?? undefined;
+
+    this._retryOptions = {
+      retryCount: options?.fetchRetryCount ?? 3,
+      retryDelay: options?.fetchRetryDelay ?? 3000,
+      retryCheck: isLangfuseFetchError,
+    };
+    this.requestTimeout = options?.requestTimeout ?? 10000; // 10 seconds
+  }
+
+  protected getCommonEventProperties(): any {
+    return {
+      $lib: this.getLibraryId(),
+      $lib_version: this.getLibraryVersion(),
+    };
+  }
+
+  on(event: string, cb: (...args: any[]) => void): () => void {
+    return this._events.on(event, cb);
+  }
+
+  debug(enabled: boolean = true): void {
+    this.removeDebugCallback?.();
+
+    this.debugMode = enabled;
+
+    if (enabled) {
+      this.removeDebugCallback = this.on("*", (event, payload) => console.log("Langfuse Debug", event, payload));
+    }
+  }
+
+  /***
+   *** Handlers for each object type
+   ***/
+  protected traceStateless(body: CreateLangfuseTraceBody): string {
+    const { id: bodyId, release: bodyRelease, ...rest } = body;
+
+    const id = bodyId ?? generateUUID();
+    const release = bodyRelease ?? this.release;
+
+    const parsedBody: CreateLangfuseTraceBody = {
+      id,
+      release,
+      ...rest,
+    };
+    this.enqueue("createTrace", parsedBody);
+    return id;
+  }
+
+  protected eventStateless(body: CreateLangfuseEventBody): string {
+    const { id: bodyId, startTime: bodyStartTime, ...rest } = body;
+
+    const id = bodyId ?? generateUUID();
+
+    const parsedBody: CreateLangfuseEventBody = {
+      id,
+      startTime: bodyStartTime ?? new Date(),
+      ...rest,
+    };
+    this.enqueue("createEvent", parsedBody);
+    return id;
+  }
+
+  protected spanStateless(body: CreateLangfuseSpanBody): string {
+    const { id: bodyId, startTime: bodyStartTime, ...rest } = body;
+
+    const id = bodyId || generateUUID();
+
+    const parsedBody: CreateLangfuseSpanBody = {
+      id,
+      startTime: bodyStartTime ?? new Date(),
+      ...rest,
+    };
+    this.enqueue("createSpan", parsedBody);
+    return id;
+  }
+
+  protected generationStateless(body: CreateLangfuseGenerationBody): string {
+    const { id: bodyId, startTime: bodyStartTime, ...rest } = body;
+
+    const id = bodyId || generateUUID();
+
+    const parsedBody: CreateLangfuseGenerationBody = {
+      id,
+      startTime: bodyStartTime ?? new Date(),
+      ...rest,
+    };
+    this.enqueue("createGeneration", parsedBody);
+    return id;
+  }
+
+  protected scoreStateless(body: CreateLangfuseScoreBody): string {
+    const { id: bodyId, ...rest } = body;
+
+    const id = bodyId || generateUUID();
+
+    const parsedBody: CreateLangfuseScoreBody = {
+      id,
+      ...rest,
+    };
+    this.enqueue("createScore", parsedBody);
+    return id;
+  }
+
+  protected updateSpanStateless(body: UpdateLangfuseSpanBody): string {
+    this.enqueue("updateSpan", body);
+    return body.spanId;
+  }
+
+  protected updateGenerationStateless(body: UpdateLangfuseGenerationBody): string {
+    this.enqueue("updateGeneration", body);
+    return body.generationId;
+  }
+
+  protected async _getDataset(name: GetLangfuseDatasetParams["datasetName"]): Promise<GetLangfuseDatasetResponse> {
+    return this.fetch(`${this.baseUrl}/api/public/datasets/${name}`, this.getFetchOptions({ method: "GET" })).then(
+      (res) => res.json()
+    );
+  }
+
+  async getDatasetRun(params: GetLangfuseDatasetRunParams): Promise<GetLangfuseDatasetRunResponse> {
+    return this.fetch(
+      `${this.baseUrl}/api/public/datasets/${params.datasetName}/runs/${params.runName}`,
+      this.getFetchOptions({ method: "GET" })
+    ).then((res) => res.json());
+  }
+
+  async createDatasetRunItem(body: CreateLangfuseDatasetRunItemBody): Promise<CreateLangfuseDatasetRunItemResponse> {
+    return this.fetch(
+      `${this.baseUrl}/api/public/dataset-run-items`,
+      this.getFetchOptions({ method: "POST", body: JSON.stringify(body) })
+    ).then((res) => res.json());
+  }
+
+  async createDataset(name: string): Promise<CreateLangfuseDatasetResponse> {
+    const body: CreateLangfuseDatasetBody = { name };
+    return this.fetch(
+      `${this.baseUrl}/api/public/datasets`,
+      this.getFetchOptions({ method: "POST", body: JSON.stringify(body) })
+    ).then((res) => res.json());
+  }
+
+  async createDatasetItem(body: CreateLangfuseDatasetItemBody): Promise<CreateLangfuseDatasetItemResponse> {
+    return this.fetch(
+      `${this.baseUrl}/api/public/dataset-items`,
+      this.getFetchOptions({ method: "POST", body: JSON.stringify(body) })
+    ).then((res) => res.json());
+  }
+
+  protected _parsePayload(response: any): any {
+    try {
+      return JSON.parse(response);
+    } catch {
+      return response;
+    }
+  }
+
+  /***
+   *** QUEUEING AND FLUSHING
+   ***/
+  protected enqueue(type: LangfuseObject, body: any): void {
+    const queue = this.getPersistedProperty<LangfuseQueueItem[]>(LangfusePersistedProperty.Queue) || [];
+
+    const id = generateUUID();
+
+    queue.push({
+      id,
+      method: LangfusePostApiRoutes[type][0],
+      apiRoute: LangfusePostApiRoutes[type][1],
+      body,
+    });
+    this.setPersistedProperty<LangfuseQueueItem[]>(LangfusePersistedProperty.Queue, queue);
+
+    this._events.emit(type, body);
+
+    // Flush queued events if we meet the flushAt length
+    if (queue.length >= this.flushAt) {
+      this.flush();
+    }
+
+    if (this.flushInterval && !this._flushTimer) {
+      this._flushTimer = safeSetTimeout(() => this.flush(), this.flushInterval);
+    }
+  }
+
+  flushAsync(): Promise<any> {
+    return Promise.all(this.flush());
+  }
+
+  // Flushes all events that are not yet sent to the server
+  // @returns {Promise[]} - list of promises for each item in the queue that is flushed
+  flush(): Promise<LangfuseFetchResponse>[] {
+    if (this._flushTimer) {
+      clearTimeout(this._flushTimer);
+      this._flushTimer = null;
+    }
+
+    const queue = this.getPersistedProperty<LangfuseQueueItem[]>(LangfusePersistedProperty.Queue) || [];
+
+    if (!queue.length) {
+      return [];
+    }
+
+    // Flush all items in queue, could also use flushAt with splice to flush only a certain number of items (e.g. when batching)
+    const items = queue;
+    this.setPersistedProperty<LangfuseQueueItem[]>(LangfusePersistedProperty.Queue, []);
+
+    // TODO: add /batch endpoint to ingest multiple events at once
+    const promises = items.map((item) => {
+      const done = (err?: any): void => {
+        if (err) {
+          this._events.emit("error", err);
+        }
+        // remove promise from pendingPromises
+        delete this.pendingPromises[item.id];
+        this._events.emit("flush", item);
+      };
+      const payload = JSON.stringify(item.body); // implicit conversion also of dates to strings
+      const url = `${this.baseUrl}${item.apiRoute}`;
+
+      const fetchOptions = this.getFetchOptions({
+        method: item.method,
+        body: payload,
+      });
+
+      const requestPromise = this.fetchWithRetry(url, fetchOptions);
+      this.pendingPromises[item.id] = requestPromise;
+      requestPromise
+        .then(() => done())
+        .catch((err) => {
+          done(err);
+        });
+      return requestPromise;
+    });
+
+    return promises;
+  }
+
+  private getFetchOptions(p: {
+    method: LangfuseFetchOptions["method"];
+    body?: LangfuseFetchOptions["body"];
+  }): LangfuseFetchOptions {
+    const fetchOptions: LangfuseFetchOptions = {
+      method: p.method,
+      headers: {
+        "Content-Type": "application/json",
+        "X-Langfuse-Sdk-Name": "langfuse-js",
+        "X-Langfuse-Sdk-Version": this.getLibraryVersion(),
+        "X-Langfuse-Sdk-Variant": this.getLibraryId(),
+        "X-Langfuse-Public-Key": this.publicKey,
+        ...this.constructAuthorizationHeader(this.publicKey, this.secretKey),
+      },
+      body: p.body,
+    };
+
+    return fetchOptions;
+  }
+
+  private constructAuthorizationHeader(
+    publicKey: string,
+    secretKey?: string
+  ): {
+    Authorization: string;
+  } {
+    if (secretKey === undefined) {
+      return { Authorization: "Bearer " + publicKey };
+    } else {
+      const encodedCredentials =
+        typeof btoa === "function"
+          ? // btoa() is available, the code is running in a browser or edge environment
+            btoa(publicKey + ":" + secretKey)
+          : // btoa() is not available, the code is running in Node.js
+            Buffer.from(publicKey + ":" + secretKey).toString("base64");
+
+      return { Authorization: "Basic " + encodedCredentials };
+    }
+  }
+
+  private async fetchWithRetry(
+    url: string,
+    options: LangfuseFetchOptions,
+    retryOptions?: RetriableOptions
+  ): Promise<LangfuseFetchResponse> {
+    (AbortSignal as any).timeout ??= function timeout(ms: number) {
+      const ctrl = new AbortController();
+      setTimeout(() => ctrl.abort(), ms);
+      return ctrl.signal;
+    };
+
+    return await retriable(
+      async () => {
+        let res: LangfuseFetchResponse | null = null;
+        try {
+          res = await this.fetch(url, {
+            signal: (AbortSignal as any).timeout(this.requestTimeout),
+            ...options,
+          });
+        } catch (e) {
+          // fetch will only throw on network errors or on timeouts
+          throw new LangfuseFetchNetworkError(e);
+        }
+        if (res.status < 200 || res.status >= 400) {
+          throw new LangfuseFetchHttpError(res);
+        }
+        return res;
+      },
+      { ...this._retryOptions, ...retryOptions },
+      (string) => this._events.emit("retry", string + ", " + url + ", " + JSON.stringify(options))
+    );
+  }
+
+  async shutdownAsync(): Promise<void> {
+    clearTimeout(this._flushTimer);
+    try {
+      await this.flushAsync();
+      await Promise.all(
+        Object.values(this.pendingPromises).map((x) =>
+          x.catch(() => {
+            // ignore errors as we are shutting down and can't deal with them anyways.
+          })
+        )
+      );
+    } catch (e) {
+      if (!isLangfuseFetchError(e)) {
+        throw e;
+      }
+      console.error("Error while shutting down Langfuse", e);
+    }
+  }
+
+  shutdown(): void {
+    void this.shutdownAsync();
+  }
+
+  protected async awaitAllQueuedAndPendingRequests(): Promise<void> {
+    clearTimeout(this._flushTimer);
+    await this.flushAsync();
+    await Promise.all(Object.values(this.pendingPromises));
+  }
+}
+
+export abstract class LangfuseWebStateless extends LangfuseCoreStateless {
+  constructor(params: { publicKey: string } & LangfuseCoreOptions) {
+    const { flushAt, flushInterval, ...rest } = params;
+    super({
+      ...rest,
+      flushAt: flushAt ?? 1,
+      flushInterval: flushInterval ?? 0,
+    });
+  }
+
+  async score(body: CreateLangfuseScoreBody): Promise<this> {
+    this.scoreStateless(body);
+    await this.awaitAllQueuedAndPendingRequests();
+    return this;
+  }
+}
+
+export abstract class LangfuseCore extends LangfuseCoreStateless {
+  constructor(params: { publicKey: string; secretKey: string } & LangfuseCoreOptions) {
+    assert(params.publicKey, "You must pass your Langfuse project's api public key.");
+    assert(params.secretKey, "You must pass your Langfuse project's api secret key.");
+    super(params);
+  }
+
+  trace(body?: CreateLangfuseTraceBody): LangfuseTraceClient {
+    const id = this.traceStateless(body ?? {});
+    const t = new LangfuseTraceClient(this, id);
+    if (getEnv("DEFER") && body) {
+      try {
+        const deferRuntime = getEnv<DeferRuntime>("__deferRuntime");
+        if (deferRuntime) {
+          deferRuntime.langfuseTraces([
+            {
+              id: id,
+              name: body.name || "",
+              url: t.getTraceUrl(),
+            },
+          ]);
+        }
+      } catch {}
+    }
+    return t;
+  }
+
+  span(body: CreateLangfuseSpanBody): LangfuseSpanClient {
+    const traceId = body.traceId || this.traceStateless({ name: body.name });
+    const id = this.spanStateless({ ...body, traceId });
+    return new LangfuseSpanClient(this, id, traceId);
+  }
+
+  generation(body: CreateLangfuseGenerationBody): LangfuseGenerationClient {
+    const traceId = body.traceId || this.traceStateless({ name: body.name });
+    const id = this.generationStateless({ ...body, traceId });
+    return new LangfuseGenerationClient(this, id, traceId);
+  }
+
+  event(body: CreateLangfuseEventBody): LangfuseEventClient {
+    const traceId = body.traceId || this.traceStateless({ name: body.name });
+    const id = this.eventStateless({ ...body, traceId });
+    return new LangfuseEventClient(this, id, traceId);
+  }
+
+  score(body: CreateLangfuseScoreBody): this {
+    this.scoreStateless(body);
+    return this;
+  }
+
+  async getDataset(name: string): Promise<{
+    id: string;
+    name: string;
+    projectId: string;
+    items: Array<{
+      id: string;
+      input: any;
+      expectedOutput?: any;
+      sourceObservationId?: string | null;
+      link: (obj: LangfuseObservationClient, runName: string) => Promise<{ id: string }>;
+    }>;
+  }> {
+    const { items, ...dataset } = await this._getDataset(name);
+
+    const returnDataset = {
+      ...dataset,
+      items: items.map((item) => ({
+        ...item,
+        link: async (obj: LangfuseObservationClient, runName: string) => {
+          await this.awaitAllQueuedAndPendingRequests();
+          const data = await this.createDatasetRunItem({
+            runName,
+            datasetItemId: item.id,
+            observationId: obj.id,
+          });
+          return data;
+        },
+      })),
+    };
+
+    return returnDataset;
+  }
+
+  _updateSpan(body: UpdateLangfuseSpanBody): this {
+    this.updateSpanStateless(body);
+    return this;
+  }
+
+  _updateGeneration(body: UpdateLangfuseGenerationBody): this {
+    this.updateGenerationStateless(body);
+    return this;
+  }
+}
+
+export abstract class LangfuseObjectClient {
+  public readonly client: LangfuseCore;
+  public readonly id: string;
+  public readonly traceId: string;
+  public readonly observationId: string | null;
+
+  constructor({
+    client,
+    id,
+    traceId,
+    observationId,
+  }: {
+    client: LangfuseCore;
+    id: string;
+    traceId: string;
+    observationId: string | null;
+  }) {
+    this.client = client;
+    this.id = id;
+    this.traceId = traceId;
+    this.observationId = observationId;
+  }
+
+  event(body: Omit<CreateLangfuseEventBody, "traceId" | "parentObservationId">): LangfuseEventClient {
+    return this.client.event({
+      ...body,
+      traceId: this.traceId,
+      parentObservationId: this.observationId,
+    });
+  }
+
+  span(body: Omit<CreateLangfuseSpanBody, "traceId" | "parentObservationId">): LangfuseSpanClient {
+    return this.client.span({
+      ...body,
+      traceId: this.traceId,
+      parentObservationId: this.observationId,
+    });
+  }
+
+  generation(body: Omit<CreateLangfuseGenerationBody, "traceId" | "parentObservationId">): LangfuseGenerationClient {
+    return this.client.generation({
+      ...body,
+      traceId: this.traceId,
+      parentObservationId: this.observationId,
+    });
+  }
+
+  score(body: Omit<CreateLangfuseScoreBody, "traceId" | "parentObservationId">): this {
+    this.client.score({
+      ...body,
+      traceId: this.traceId,
+      observationId: this.observationId,
+    });
+    return this;
+  }
+
+  getTraceUrl(): string {
+    return `${this.client.baseUrl}/trace/${this.traceId}`;
+  }
+}
+
+export class LangfuseTraceClient extends LangfuseObjectClient {
+  constructor(client: LangfuseCore, traceId: string) {
+    super({ client, id: traceId, traceId, observationId: null });
+  }
+
+  update(body: Omit<CreateLangfuseTraceBody, "id">): this {
+    this.client.trace({
+      ...body,
+      id: this.id,
+    });
+    return this;
+  }
+}
+
+abstract class LangfuseObservationClient extends LangfuseObjectClient {
+  constructor(client: LangfuseCore, id: string, traceId: string) {
+    super({ client, id, traceId, observationId: id });
+  }
+}
+
+export class LangfuseSpanClient extends LangfuseObservationClient {
+  constructor(client: LangfuseCore, id: string, traceId: string) {
+    super(client, id, traceId);
+  }
+
+  update(body: Omit<UpdateLangfuseSpanBody, "spanId" | "traceId">): this {
+    this.client._updateSpan({
+      ...body,
+      spanId: this.id,
+      traceId: this.traceId,
+    });
+    return this;
+  }
+
+  end(body?: Omit<UpdateLangfuseSpanBody, "spanId" | "endTime" | "traceId">): this {
+    this.client._updateSpan({
+      ...body,
+      spanId: this.id,
+      traceId: this.traceId,
+      endTime: new Date(),
+    });
+    return this;
+  }
+}
+
+export class LangfuseGenerationClient extends LangfuseObservationClient {
+  constructor(client: LangfuseCore, id: string, traceId: string) {
+    super(client, id, traceId);
+  }
+
+  update(body: Omit<UpdateLangfuseGenerationBody, "generationId" | "traceId">): this {
+    this.client._updateGeneration({
+      ...body,
+      generationId: this.id,
+      traceId: this.traceId,
+    });
+    return this;
+  }
+
+  end(body?: Omit<UpdateLangfuseGenerationBody, "generationId" | "traceId" | "endTime">): this {
+    this.client._updateGeneration({
+      ...body,
+      generationId: this.id,
+      traceId: this.traceId,
+      endTime: new Date(),
+    });
+    return this;
+  }
+}
+
+export class LangfuseEventClient extends LangfuseObservationClient {
+  constructor(client: LangfuseCore, id: string, traceId: string) {
+    super(client, id, traceId);
+  }
+}
+
+export * from "./types";
+export * from "./openapi/server";
