+++ package.json
@@ -66,7 +66,8 @@
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "tslib": "^2.6.1",
-    "typescript": "^5.1.6"
+    "typescript": "^5.1.6",
+    "yalc": "^1.0.0-pre.53"
   },
   "husky": {
     "hooks": {
@@ -75,8 +76,5 @@
   },
   "lint-staged": {
     "*.{ts,js,css,scss}": "prettier --write"
-  },
-  "dependencies": {
-    "yalc": "^1.0.0-pre.53"
   }
 }
